requirements.txt
Django==3.0.6
WTForms==2.2.1
#PyMySQL==0.9.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
pip==19.1.1
pytz==2020.1
requests==2.23.0
setuptools==46.2.0
sqlparse==0.3.1
urllib3==1.25.9
psycopg2==2.7.5
gunicorn==20.0.4
django-heroku
psycopg2-binary
Unidecode==1.1.1

redis==3.5.1
Pillow==7.1.2
PyPDF2==1.26.0
asgiref==3.2.7
blinker==1.4
public==2019.4.13
pycryptodomex==3.9.7
values==2019.4.13